move tests to separate file, and make them pass
